universal_ai_agent/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
sounddevice
speechrecognition
pillow
transformers
sentencepiece
torch
langchain
pyttsx3
pytest
langdetect
streamlit
